refactor(swaggers): moved swagger to config
