+++ src/main/config/index.ts
@@ -1,1 +1,2 @@
 export * from './environments'
+export * from './swagger'
